ETL_project_end_to_end/requirements.txt
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
openpyxl==3.1.2
xlrd==2.0.1
jinja2==3.1.2
tabulate==0.9.0
tqdm==4.66.1
colorama==0.4.6 